requirements.txt
PyYAML
adafruit-circuitpython-ssd1306
Pillow
requests
